match_backend/requirements.txt
# FastAPI 웹 프레임워크 (메인 프론트엔드)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 템플릿 엔진
jinja2>=3.1.0

# 폼 처리
python-multipart>=0.0.6

# Flask 제거 - FastAPI로 완전 전환

# OpenAI API 클라이언트
openai>=1.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 데이터 검증
pydantic>=2.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 텍스트 처리 (선택사항)
nltk>=3.8.0
scikit-learn>=1.3.0

# 기타 유틸리티
requests>=2.31.0
python-dateutil>=2.8.0
